Rename modules to match PEP 8 recommendations
https://peps.python.org/pep-0008/#package-and-module-names
